Bump linkerd-validator to v0.1.1
A new release of linkerd-validator is available, v0.1.1. The new version
includes better logging to allow users to determine whether failures
occur as a result of their environment or the tool itself.

To use the new version in linkerd, we bump the version of the binary in
the proxy dockerfile.

Signed-off-by: Matei David <matei@buoyant.io>