Fix newline in destination and identity templates
When using the normal initContainer (`proxy-init`) it seems that the
template renders a whitespace between the initContainer field name and
the start of the init container section (i.e `- args`). It seems that
trimming spaces from preceding text, and trimming spaces from succeeding
text in the comment seems to fix it.

Signed-off-by: Matei David <matei@buoyant.io>