dev: Always rebuild the CLI in bin/linkerd
`bin/linkerd` may use an outdated cached binary. There's no mechanism to
invalidate this cache except to manually delete the file. This is
cumbersome, and a common source of developer confusion.

This change removes this caching logic so that the binary is always
rebuilt. This is relatively quick (<2s) when nothing has changed.