Refactor `linkerd check` calls in the integration tests
Extracted logic into the new file `testutil/test_helper_check.go` which exposes the functions `TestCheckPre`, `TestCheck` and `TestCheckProxy`.

`linkerd check --output json` is called so its output is properly captured without the need of golden files.

Besides checking that there are no errors (although warnings are allowed), we check that the expected check categories are returned.

The plan is to leverage this in #9856 when re-enabling the helm-upgrade test.