policy: include colliding Server in validation error
Currently, when the policy controller's validating admission webhook
rejects a Server because it collides with an existing one, it's
difficult to determine which resource the new Server would collide with
(see #10153). Therefore, we should update the error message to include
the existing Server.

Since the two Servers are only considered colliding if they have the
same name and are in the same namespace, just including the name and
namespace of the existing Server in the error isn't very helpful.
Therefore, I've changed the returned error to include the entire YAML
representation of the existing Server. Hopefully this isn't too verbose
--- I'm open to other suggestions if it is.

Fixes #10153