Protect against nil dereference in resources helm template (#10235)
Fixes #10138 

Evaluating Helm expressions like `.cpu.limit` will fail with a nil pointer dereference error if `.cpu` is nil.  If, for example, `.memory` is set but `.cpu` is not, the resources template will be executed but will fail.

We add parentheses to cause these expressions to be evaluated as a pipeline.  If the input to a pipeline stage is an empty value (such as nil), no output will be emitted to the next stage of the pipeline.  This allows for more graceful dereference chaining.  For example, when evaluating `(.cpu).limit`, if `(.cpu)` is nil, the rendering engine will not try to evaluate `nil.limit` but instead will emit no output for this expression.

Signed-off-by: Alex Leong <alex@buoyant.io>