Don't use pods in map keys in server watcher (#10245)
Fixes #10205

The server watches uses `*corev1.Pod` in the key of it's subscriber map which is potentially brittle since it's not obvious what exactly should count as equality for full pod resources.

More robust is to use a unique per-pod identifier in the key: the pod's name and namespace.  We therefore more the pod resource into the value type of the map since we still need it when we receive a Server update so that we can iterate through all subscriptions and find pods which match the Server's pod selector. 

Signed-off-by: Alex Leong <alex@buoyant.io>