Add support to cross-build policy controller
The policy controller is currently built by cross-building, however, the
process is not very robust. In order to cross-build, we have to build
each image on the host platform (typically amd64), install a target
architecture compiler, and use a distroless runtime image tied to the
target arch. Furthermore, arm targets also have to compile using rustls
instead of openssl. The result is that we have three different
dockerfiles that build the image: each one specialized for a target
arch. At the end, the manifests are combined to provide a multi-arch
image.

By using a different base layer for the builder (dev container), we can
cross compile more efficiently. First, instead of dynamically linking
the TLS library, we statically link openssl. This is done by vendoring
the dependency and feature gating it (available as a default feature but
can be turned off to compile with rustls). This is all abstracted away
by using a `just` script present on the base image; the recipe sets up
the environment. Second, since all binaries are compiled using static
linking, we can use an architecture agnostic `scratch` runtime layer.

This greatly simplifies the process. We only need to maintain one docker
file. Details about the compilation (which compiler to use) are
abstracted away in the `just` file. The image includes dependencies to
build for all three target architectures supported by Linkerd: amd64,
arm and arm64 (clang, libssl). All targets are now using openssl.

Signed-off-by: Matei David <matei@buoyant.io>