Update Helm chart docs with external issuer info (#10349)
Visibility into our documentation should be improved around the usage of
an external CA and/or an external issuer. To some of our users it is not
clear that an external issuer, for example, may be used, by changing the
identity issuer's scheme.

To make it more visible, this PR updates the Helm chart docs for
`linkerd-control-plane`. A short explainer is given on how Linkerd may
be installed without explicitly passing in any certificate resources. A
more comprehensive explanation is linked (our docs on Automating
Certificate Rotation should serve as a more thorough walkthrough).

Closes #10229

Signed-off-by: Matei David <matei@buoyant.io>

---------

Signed-off-by: Matei David <matei@buoyant.io>