Problem: Namespaces and services passed to are annotated properly but print skipped to stderr.
This fix makes sure that changes to when transforming input are appropriately applied to the resulting Go slice. Previously, changes were not being properly propogated to the slice: this change makes sure that changes to the field are properly made. Returning the slice inline makes it explicit that just one is being returned in the slice and that the changes to are being passed. Unit test cases are also changed.

Fixes #10231

For example testing, run:

Signed-off-by: Scott Tran <scotttran@utexas.edu>