Add support for customizable ignore headers list in Linkerd Tap
Currently, Linkerd Tap displays all headers by default, which can contain irrelevant or sensitive information that users may want to exclude from being displayed. Therefore, there is a need for a feature that allows users to set their own ignore headers list, so that they can choose which headers to exclude from the output of Linkerd Tap on the dashboard.

This commit adds a new helm value, `tap.ignoredHeaders`, to the Linkerd-viz chart. This value allows users to specify a comma-separated list of headers to be ignored by Linkerd Tap. The default list includes commonly irrelevant and sensitive headers such as authorization tokens and cookies.

Once merged, every comma-seperated value in `tap.ignoredHeaders` will be removed from the headers in the Tap Dashboard when expanding a request headers result.

Fixes #10389

Signed-off-by: Ryan Hristovski <ryan.hristovski@docker.com>

---

This is my first time contributing to this repo so I am sure refinement or even a different approach may be suggested which I'm open to.

I understand `tap` itself will need to be updated as well which I can do if this gets a positive signal. As of right now the headers will still be visible when running the cli