Add lease claims to status controller (#10424)
This adds lease claims to the policy status controller so that upon startup, a
status controller attempts to claim the `status-controller` lease in the
`linkerd` namespace. With this lease, we can enforce leader election and ensure
that only one status controller on a cluster is attempting to patch HTTPRoute’s
`status` field.

Upon startup, the status controller now attempts to create the
`status-controller` lease — it will handle failure if the lease is already
present on the cluster. It then spawns a task for attempting to claim this lease
and sends all claim updates to the index `Index`.

Currently, `Index.claims` is not used, but in follow-up changes we can check
against the current claim for determining if the status controller is the
current leader on the cluster. If it is, we can make decisions about sending
updates or not to the controller `Controller`.

### Testing
Currently I’ve only manually tested this, but integration tests will definitely
be helpful follow-ups. For manually testing, I’ve asserted that the
`status-controller` is claimed when one or more status controllers startup and
are running on a cluster. I’ve also asserted that when the current leader is
deleted, another status controller claims the lease. Below is the summary of how
I tested it

```shell
$ linkerd install --ha |kubectl apply -f -
…

$ kubectl get -n linkerd leases status-controller
NAME                HOLDER                                 AGE
status-controller   linkerd-destination-747b456876-dcwlb   15h

$ kubectl delete -n linkerd pod linkerd-destination-747b456876-dcwlb
pod "linkerd-destination-747b456876-dcwlb" deleted

$ kubectl get -n linkerd leases status-controller
NAME                HOLDER                                 AGE
status-controller   linkerd-destination-747b456876-5zpwd   15h
```

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>