Consider status-controller lease holder when patching HTTPRoutes (#10444)
https://github.com/linkerd/linkerd2/pull/10424 added the status-controller lease
used by the policy status controller. When the status controller starts up, it
tries to become the holder of that lease; if it fails to become the holder
because another status controller already holds the lease, it keeps trying to
become the holder for its lifetime.

This change continues on that work and now considers the holder of the
status-controller lease when applying patches to HTTPRoutes. If a status
controller is not the holder of the lease, its `Index` does not send patch
updates to its `Controller`; it does keep the index up-to-date though so that if
it does become the holder in the future it has a correct view of the cluster.

Additionally, the status controller `Controller` will now attempt to reconcile
the cluster periodically. This is to cover for any failure scenarios when
applying a patch fails.

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>