Consider HTTPRoute `status` in policy controller (#10486)
When the policy controller is creating the inbound server which contains the
HTTPRoutes that a policy client should consider, we now consider the `status`
field and filter out HTTPRoutes which have not been accepted by the status
controller.

When creating the HTTPRoutes for an inbound server, the policy controller now
looks at the `status` field for each HTTPRoute. It filters out statuses which
are not from the policy status controller. For statuses which are from the
policy status controller, it ensures that it has been accepted by a Server on
the cluster.

The tests have been updated in the index; these were failing without an
`Accepted` status.

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>