proxy-injector: Skip Kube API server ports
The destination and identity controllers skip ports that are used by the
kubernetes API server to avoid startup ordering issues and to ensure
that the controllers are able to communicate with the kubernetes API
without requiring the control plane for discovery.

To avoid startup issues in the proxy-injector, the same configuration
should be applied.