ci: increase timeout for `build-core`
Policy controller docker builds seem to keep hitting the 15-minute
timeout for this job (see e.g. [this comment][1]). We should probably
try to figure out if there's anything we can do to make that build
faster, but for now, this PR bumps the timeout up by 5 minutes.
Hopefully this is enough to get CI passing reliably, although we could
bump it up by 10 if it's not...

[1]: https://github.com/linkerd/linkerd2/pull/10547#issuecomment-1472393611