Consider only running pods when sending `Add` events in the discovery API
When a pod has entered a `Failed` phase (for whatever reason), the API
Server does not deregister it, and its spec is still available for human
and machine consumption. Typically, its IP address will still be
available in the status. As a result, we may end up with duplicate IP
addresses being sent as part of destination updates. Sending a pod that
is not able to receive traffic is erroneous and can often break the
observed metrics and edges between endpoints.

To fix this, this change introduces a check in the endpoint translator
where an address' pod is checked for readiness before being included
in the total number of endpoints available for a given port. This should
eliminate any duplication. On the `Remove` side, attempting to delete an
inexistent ID from the cache will result in a noop. This ensures that a
pod may still be deleted when it transitions into a Failed state (since it was
added when it was running).

Fixes #10562

Signed-off-by: Matei David <matei@buoyant.io>