Avoid busy looping after losing lease watch (#10584)
Transient API errors can cause the policy controller's claim watch to die (see olix0r/kubert#134). In this case, the policy controller enters a busy loop and consumes large amounts of CPU.

Ideally, the Kubert lease module should handle errors internally and present an infallible interface. In absence of that, it's better for us to panic when we lose the watch and let Kubernetes restart the container, rather than busy looping.

Signed-off-by: Alex Leong <alex@buoyant.io>
Co-authored-by: Oliver Gould <ver@buoyant.io>